modify image check

+++ scripts/check_images.sh
@@ -23,10 +23,16 @@ find content -name "*.md" | while read -r file; do
             continue
         fi
 
+        if [[ "$img_path" == https://* ]]; then
+            echo "ERROR: $img_path is external URL in '$file'"
+            echo "1" > "$temp_error_file"
+            continue
+        fi
+
         cd "$(dirname "$file")"
 
         # echo "Checking image: $img_path"
-        if [[ ! -f "$img_path" ]]; then
+        if [[ "$img_path" != https://* && -f "$img_path" ]]; then
             echo "1" > "$temp_error_file"
             echo "ERROR: '$img_path' not found in '$file'"
             continue
